Use the raw Json value for message ids.
The spec allows ids to be either Number or String. Since the server simply
carries this value around so that it can submit it with the response, the
simplest solution seems to be to pass the Json type through unexamined and send
it right back to the client.

Thanks to the magic of type inference, the change is isolated and nearly
trivial.